Implemented Plant class to SimulatorView with errors
Added subclasses to define the current different types of plants, and placed them in the simulatorView class to allow them to show in the GUI. Still requires adding them as a food source for respective animals. Plant superclass is sometimes called instead of the respctive subclass - needs to be fixed.